Fixing the way maestro-ng is invoked
On a previous commit, we introduced the usage of environment variables
to pass values references defined within the maestro.yaml file. These
weren't being read correctly and this commit fixes this.

The container for maestro-ng needs to be given the entry-point
explicitly. Also, the environment variables have to be defined
explicitly also (for the `docker run` command). The run.sh script was
updated to reflect this too (fixes #9).

The plot script now generates plots with error bars for standard error
instead of standard deviation (fixes #10)