Adding results from old vs. new hard drive small experiment
We executed the scalability experiment on the original hardware, up to
4 OSD nodes. We also executed the same experiment by removing the old
drives and putting new hard drives.

The old model is a 250GB Seagate Barracuda ES (ST3250620NS). New hard
drives correspond to 500GB Western Digital Re (WD5003ABYZ) drives.

The reason for the change in results with respect to the first four
points of Figure 2 is the following. As mentioned earlier, Ceph issues
two I/O calls on each write request, one of them being asynchronous. The
cgroups blkio controller responsible for limiting I/O on block devices
(which we configure to 30 MB/s) cannot throttle asynchronous I/O
operations since this type of requests go to a queue that is shared at
the system level by all containers running in the host.